Convert long long and double to int64_t; Added more tests

+++ client-server/circularOrderedArray.c
@@ -5,13 +5,13 @@ void initCircularOrderedArray(CircularOrderedArray * array){
 	array->size = 0;
 }
 
-void swap(double * first, double * second){
-	double aux = *first;
+void swap(int64_t * first, int64_t * second){
+	int64_t aux = *first;
 	*first=*second;
 	*second=aux;
 }
 
-int orderRight(double * array, int position, int size){
+int orderRight(int64_t * array, int position, int size){
 	int nextPosition = position;
 	while((nextPosition + 1 < size) && array[nextPosition] > array[nextPosition+1]){ //hasNext && bigger than next
 		swap(&array[nextPosition], &array[nextPosition+1]);
@@ -20,27 +20,27 @@ int orderRight(double * array, int position, int size){
 	return position!=nextPosition;
 }
 
-void orderLeft(double * array, int position){
+void orderLeft(int64_t * array, int position){
 	while(position > 0 && array[position] < array[position-1]){ //hasPrevious && smaller than previous
 		swap(&array[position], &array[position-1]);
 		position--;
 	}
 }
 
-void orderPosition(double * array, int position, int size){
+void orderPosition(int64_t * array, int position, int size){
 	if(!orderRight(array, position, size)){
 		orderLeft(array, position);
 	}	
 }
 
-void insertOrdered(CircularOrderedArray* array, double value){
+void insertOrdered(CircularOrderedArray* array, int64_t value){
 	int insertPosition;
 	if(array->size < CIRCULAR_ORDERED_ARRAY_MAX_SIZE){
 		insertPosition = array->size;
 		array->fifo[array->size]=value;
 		array->size++;
 	} else {
-		double valueToRemove = array->fifo[array->next];
+		int64_t valueToRemove = array->fifo[array->next];
 		for(int i = 0; i< CIRCULAR_ORDERED_ARRAY_MAX_SIZE; i++){
 			if(array->array[i] == valueToRemove){
 				insertPosition = i;
@@ -55,6 +55,6 @@ void insertOrdered(CircularOrderedArray* array, double value){
 	orderPosition(array->array, insertPosition, array->size);
 }
 
-double median(CircularOrderedArray* array){
+int64_t median(CircularOrderedArray* array){
 	return array->array[array->size/2];
 }
